Refactor namings.
